# docs/openapi.yaml
openapi: 3.0.3
info:
  title: European Energy Prices API
  description: |
    A comprehensive REST API for fetching real-time electricity prices across 9 European countries, 
    with support for energy provider markups, VAT calculations, and flexible time periods.
    
    ## Features
    - **9 European Countries** - Netherlands, Germany, Belgium, France, Austria, Switzerland, Denmark, Norway, Sweden
    - **Multiple Time Periods** - Today, tomorrow, next 24 hours, custom date ranges
    - **Provider Markups** - Built-in support for energy provider pricing
    - **Flexible VAT** - Country-specific VAT rates or custom percentages
    - **Multi-Currency** - EUR, CHF, DKK, NOK, SEK support
    - **Timezone Aware** - Proper timezone handling for each country
    
    ## Data Sources
    - **Primary**: Stekker.app - European energy price aggregator
    - **Backup**: ENTSOE Transparency Platform - Official EU energy data
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/your-username/european-energy-prices-api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-api-domain.com
    description: Production server
...
paths:
...
  /api/{country}/today:
    get:
      tags:
        - Countries
      summary: Get today's energy prices
      description: Get energy prices for a specific country from 00:00 to 23:59 of the current day
      parameters:
        - $ref: '#/components/parameters/CountryPath'
        - $ref: '#/components/parameters/FixedMarkup'
        - $ref: '#/components/parameters/VariableMarkup'
        - $ref: '#/components/parameters/VAT'
        - $ref: '#/components/parameters/AutoVAT'
        - $ref: '#/components/parameters/RoundTo'
      responses:
        '200':
          description: Today's energy prices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/{country}/tomorrow:
    get:
      tags:
        - Countries
      summary: Get tomorrow's energy prices
      description: Get energy prices for a specific country from 00:00 to 23:59 of the next day
      parameters:
        - $ref: '#/components/parameters/CountryPath'
        - $ref: '#/components/parameters/FixedMarkup'
        - $ref: '#/components/parameters/VariableMarkup'
        - $ref: '#/components/parameters/VAT'
        - $ref: '#/components/parameters/AutoVAT'
        - $ref: '#/components/parameters/RoundTo'
      responses:
        '200':
          description: Tomorrow's energy prices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/{country}/next24h:
    get:
      tags:
        - Countries
      summary: Get next 24 hours of energy prices
      description: Get energy prices for the next 24 hours from the current time, spanning across days if necessary
      parameters:
        - $ref: '#/components/parameters/CountryPath'
        - $ref: '#/components/parameters/FixedMarkup'
        - $ref: '#/components/parameters/VariableMarkup'
        - $ref: '#/components/parameters/VAT'
        - $ref: '#/components/parameters/AutoVAT'
        - $ref: '#/components/parameters/RoundTo'
      responses:
        '200':
          description: Next 24 hours of energy prices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Next24HoursResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    CountryPath:
      name: country
      in: path
      required: true
      description: |
        Country code (ISO 3166-1 alpha-2). 
        Supported: nl, de, be, fr, at, ch, dk, no, se
      schema:
        type: string
        pattern: '^[a-z]{2}$'
        example: "nl"
    FixedMarkup:
      name: markup
      in: query
      description: Fixed markup per kWh in local currency (e.g., 0.024 for €0.024/kWh)
      schema:
        type: number
        format: float
        minimum: 0
        example: 0.024
    VariableMarkup:
      name: variableMarkup
      in: query
      description: Variable markup as percentage (e.g., 5 for 5%)
      schema:
        type: number
        format: float
        minimum: 0
        example: 5
    VAT:
      name: vat
      in: query
      description: VAT rate as decimal (e.g., 0.21 for 21%)
      schema:
        type: number
        format: float
        minimum: 0
        maximum: 1
        example: 0.21
    AutoVAT:
      name: autoVat
      in: query
      description: Use country's default VAT rate
      schema:
        type: boolean
        example: true
    RoundTo:
      name: roundTo
      in: query
      description: Number of decimal places to round prices to
      schema:
        type: integer
        minimum: 0
        maximum: 10
        default: 5
        example: 5
  schemas:
...
    PricesResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        country:
          $ref: '#/components/schemas/CountryInfo'
        data:
          type: array
          items:
            $ref: '#/components/schemas/PriceData'
        markup:
          $ref: '#/components/schemas/MarkupInfo'
        fetchedAt:
          type: string
          format: date-time
        info:
          $ref: '#/components/schemas/ResponseInfo'
    Next24HoursResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        country:
          $ref: '#/components/schemas/CountryInfo'
        data:
          type: array
          items:
            $ref: '#/components/schemas/ExtendedPriceData'
        markup:
          $ref: '#/components/schemas/MarkupInfo'
        fetchedAt:
          type: string
          format: date-time
        info:
          $ref: '#/components/schemas/ExtendedResponseInfo'
...
    CountryInfo:
      type: object
      properties:
        code:
          type: string
          example: "NL"
        name:
          type: string
          example: "Netherlands"
        currency:
          type: string
          example: "EUR"
        timezone:
          type: string
          example: "Europe/Amsterdam"
    PriceData:
      type: object
      properties:
        time:
          type: string
          format: date-time
          description: ISO 8601 timestamp
          example: "2025-05-23T00:00:00.000Z"
        priceRaw:
          type: number
          format: float
          description: Raw market price in local currency per MWh
          example: 87.12
        price:
          type: number
          format: float
          description: Final price in local currency per kWh (including markup and VAT)
          example: 0.13442
        hour:
          type: string
          description: Local time in HH:MM format
          example: "02:00"
        localTime:
          type: string
          description: Formatted local time
          example: "02:00"
        markup:
          $ref: '#/components/schemas/MarkupDetails'
    ExtendedPriceData:
      allOf:
        - $ref: '#/components/schemas/PriceData'
        - type: object
          properties:
            hourFromNow:
              type: integer
              description: Hours from current time
              example: 0
            dayOfWeek:
              type: string
              example: "Friday"
...
    MarkupInfo:
      type: object
      properties:
        fixedMarkup:
          type: number
          example: 0.024
        variableMarkup:
          type: number
          example: 0
        vat:
          type: number
          example: 0.21
        includeVat:
          type: boolean
          example: true
        roundTo:
          type: integer
          example: 5
    MarkupDetails:
      type: object
      properties:
        fixed:
          type: number
          example: 0.024
        variable:
          type: number
          example: 0
        vat:
          type: number
          example: 0.21
        vatPercent:
          type: string
          example: "21%"
        originalPrice:
          type: number
          example: 0.08712
...
    ResponseInfo:
      type: object
      properties:
        type:
          type: string
          example: "today"
        date:
          type: string
          example: "22-5-2025"
        totalHours:
          type: integer
          example: 24
        priceUnit:
          type: string
          example: "EUR/kWh"
        timezone:
          type: string
          example: "Europe/Amsterdam"
    ExtendedResponseInfo:
      allOf:
        - $ref: '#/components/schemas/ResponseInfo'
        - type: object
          properties:
            startTime:
              type: string
              format: date-time
              example: "2025-05-23T12:00:00.000Z"
            endTime:
              type: string
              format: date-time
              example: "2025-05-24T12:00:00.000Z"
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Unsupported country: xx. Use /api/countries to see supported countries."
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...